Safari-609.1.11.1/LayoutTests/crypto/subtle/rsaes-pkcs1-v1_5-import-jwk-private-key-expected.txt

Test importing a JWK RSAES-PKCS1-v1_5 private key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS privateKey.toString() is '[object CryptoKey]'
PASS privateKey.type is 'private'
PASS privateKey.extractable is true
PASS privateKey.algorithm.name is 'RSAES-PKCS1-v1_5'
PASS privateKey.algorithm.modulusLength is 2048
PASS bytesToHexString(privateKey.algorithm.publicExponent) is '010001'
PASS privateKey.algorithm.hash is undefined.
PASS privateKey.usages is ['decrypt']
PASS successfullyParsed is true

TEST COMPLETE
